requirements.txt
Flask~=3.0.3
connexion~=3.0.6
connexion[uvicorn]
connexion[flask]
connexion[swagger-ui]
langchain~=0.1.13
python-dotenv~=1.0.1
langchain-openai~=0.1.0
openai~=1.14.2
PyYAML~=6.0.1
requests~=2.31.0